refactor: bookmarks, accounts 복수형 패키지 이름 수정
