+++ android/app/src/main/java/com/on/turip/data/account/dto/MyProfileResponse.kt
@@ -1,0 +1,14 @@
+package com.on.turip.data.account.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MyProfileResponse(
+    @SerialName("id")
+    val id: Long,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("role")
+    val role: String,
+)
