--- android/app/src/main/java/com/on/turip/data/accounts/dto/MyProfileResponse.kt
@@ -1,14 +1,0 @@
-package com.on.turip.data.accounts.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MyProfileResponse(
-    @SerialName("id")
-    val id: Long,
-    @SerialName("nickname")
-    val nickname: String,
-    @SerialName("role")
-    val role: String,
-)
