+++ android/app/src/main/java/com/on/turip/data/bookmark/dto/BookmarkContentsResponse.kt
@@ -1,0 +1,12 @@
+package com.on.turip.data.bookmark.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BookmarkContentsResponse(
+    @SerialName("contents")
+    val contents: List<BookmarkContentResponse>,
+    @SerialName("loadable")
+    val loadable: Boolean,
+)
