+++ android/app/src/main/java/com/on/turip/data/bookmark/dto/BookmarkCountResponse.kt
@@ -1,0 +1,10 @@
+package com.on.turip.data.bookmark.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BookmarkCountResponse(
+    @SerialName("count")
+    val count: Int,
+)
