--- android/app/src/main/java/com/on/turip/data/bookmarks/dto/BookmarkAddRequest.kt
@@ -1,10 +1,0 @@
-package com.on.turip.data.bookmarks.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class BookmarkAddRequest(
-    @SerialName("contentId")
-    val contentId: Long,
-)
