+++ android/app/src/main/java/com/on/turip/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.on.turip.di
 
-import com.on.turip.data.accounts.repository.DefaultAccountRepository
-import com.on.turip.data.bookmarks.repository.DefaultBookmarkRepository
+import com.on.turip.data.account.repository.DefaultAccountRepository
+import com.on.turip.data.bookmark.repository.DefaultBookmarkRepository
 import com.on.turip.data.content.repository.DefaultContentRepository
 import com.on.turip.data.login.repository.DefaultAuthRepository
 import com.on.turip.data.login.repository.DefaultGuestRepository
@@ -10,7 +10,7 @@ import com.on.turip.data.region.repository.DefaultRegionRepository
 import com.on.turip.data.searchhistory.repository.DefaultSearchHistoryRepository
 import com.on.turip.data.turip.repository.DefaultTuripRepository
 import com.on.turip.data.userstorage.repository.DefaultUserStorageRepository
-import com.on.turip.domain.accounts.AccountRepository
+import com.on.turip.domain.account.AccountRepository
 import com.on.turip.domain.bookmark.repository.BookmarkRepository
 import com.on.turip.domain.content.repository.ContentRepository
 import com.on.turip.domain.login.AuthRepository
