+++ android/app/src/main/java/com/on/turip/di/ServiceModule.kt
@@ -1,9 +1,9 @@
 package com.on.turip.di
 
-import com.on.turip.data.accounts.service.AccountService
-import com.on.turip.data.accounts.service.createAccountService
-import com.on.turip.data.bookmarks.service.BookmarkService
-import com.on.turip.data.bookmarks.service.createBookmarkService
+import com.on.turip.data.account.service.AccountService
+import com.on.turip.data.account.service.createAccountService
+import com.on.turip.data.bookmark.service.BookmarkService
+import com.on.turip.data.bookmark.service.createBookmarkService
 import com.on.turip.data.content.service.ContentService
 import com.on.turip.data.content.service.createContentService
 import com.on.turip.data.login.service.AuthService
