+++ android/app/src/main/java/com/on/turip/domain/account/Account.kt
@@ -1,0 +1,7 @@
+package com.on.turip.domain.account
+
+data class Account(
+    val id: Long,
+    val nickname: String,
+    val role: Role,
+)
