+++ android/app/src/main/java/com/on/turip/domain/account/AccountRepository.kt
@@ -1,0 +1,7 @@
+package com.on.turip.domain.account
+
+import com.on.turip.core.result.TuripResult
+
+interface AccountRepository {
+    suspend fun loadMyProfile(): TuripResult<Account>
+}
