+++ android/app/src/main/java/com/on/turip/domain/account/Role.kt
@@ -1,0 +1,16 @@
+package com.on.turip.domain.account
+
+enum class Role(
+    private val tag: String,
+) {
+    MEMBER("MEMBER"),
+    GUEST("GUEST"),
+    ;
+
+    companion object {
+        fun from(role: String): Role {
+            val tag = role.uppercase()
+            return entries.firstOrNull { it.tag == tag } ?: GUEST
+        }
+    }
+}
