--- android/app/src/main/java/com/on/turip/domain/accounts/Account.kt
@@ -1,7 +1,0 @@
-package com.on.turip.domain.accounts
-
-data class Account(
-    val id: Long,
-    val nickname: String,
-    val role: Role,
-)
