--- android/app/src/main/java/com/on/turip/domain/accounts/AccountRepository.kt
@@ -1,7 +1,0 @@
-package com.on.turip.domain.accounts
-
-import com.on.turip.core.result.TuripResult
-
-interface AccountRepository {
-    suspend fun loadMyProfile(): TuripResult<Account>
-}
