+++ android/app/src/main/java/com/on/turip/ui/bookmarks/BookmarkContentActivity.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import com.on.turip.ui.compose.bookmarks.BookmarkContentScreen
+import com.on.turip.ui.compose.bookmark.BookmarkContentScreen
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
 import com.on.turip.ui.login.LoginActivity
 import com.on.turip.ui.trip.TripDetailActivity
