+++ android/app/src/main/java/com/on/turip/ui/compose/bookmark/BookmarkContentUiEffect.kt
@@ -1,0 +1,9 @@
+package com.on.turip.ui.compose.bookmark
+
+sealed interface BookmarkContentUiEffect {
+    data object NavigateToLogin : BookmarkContentUiEffect
+
+    data object ShowBookmarkRemoveFailed : BookmarkContentUiEffect
+
+    data object BookmarkRemoved : BookmarkContentUiEffect
+}
