+++ android/app/src/main/java/com/on/turip/ui/compose/bookmark/component/BookmarkContentItem.kt
@@ -1,0 +1,122 @@
+package com.on.turip.ui.compose.bookmark.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.tooling.preview.Preview
+import com.on.turip.domain.bookmark.BookmarkContent
+import com.on.turip.domain.content.Content
+import com.on.turip.domain.content.video.VideoData
+import com.on.turip.domain.creator.Creator
+import com.on.turip.domain.region.City
+import com.on.turip.domain.trip.TripDuration
+import com.on.turip.ui.common.component.bookmark.BookmarkContentMetaSection
+import com.on.turip.ui.common.component.bookmark.BookmarkContentTitleRow
+import com.on.turip.ui.common.component.content.ContentThumbnail
+import com.on.turip.ui.compose.designsystem.theme.TuripTheme
+
+@Composable
+fun BookmarkContentItem(
+    content: BookmarkContent,
+    onContentClick: (contentId: Long) -> Unit,
+    onRemoveBookmark: (contentId: Long) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Column(
+        modifier =
+            modifier
+                .fillMaxWidth()
+                .clip(TuripTheme.shape.container)
+                .clickable { onContentClick(content.content.id) }
+                .background(TuripTheme.colors.white)
+                .padding(TuripTheme.spacing.extraSmall),
+    ) {
+        ContentThumbnail(
+            imageUrl = content.content.videoData.url,
+            modifier = Modifier.fillMaxWidth(),
+        )
+
+        Spacer(modifier = Modifier.height(TuripTheme.spacing.medium))
+
+        BookmarkContentTitleRow(
+            title = content.content.videoData.title,
+            trailing = {
+                BookmarkRegionChip(
+                    regionName = content.content.city.name,
+                    modifier = Modifier.padding(horizontal = TuripTheme.spacing.small),
+                )
+            },
+        )
+
+        Spacer(modifier = Modifier.height(TuripTheme.spacing.small))
+
+        BookmarkContentMetaSection(
+            item = content,
+            onRemoveBookmark = onRemoveBookmark,
+        )
+    }
+}
+
+@Composable
+private fun BookmarkRegionChip(
+    regionName: String,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        modifier =
+            modifier
+                .wrapContentSize()
+                .background(
+                    color = TuripTheme.colors.chipBackground,
+                    shape = TuripTheme.shape.chip,
+                ).padding(
+                    horizontal = TuripTheme.spacing.medium,
+                    vertical = TuripTheme.spacing.extraSmall,
+                ),
+    ) {
+        Text(
+            text = regionName,
+            style = TuripTheme.typography.info1,
+            color = TuripTheme.colors.black,
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun BookmarkContentItemPreview() {
+    TuripTheme {
+        val content =
+            BookmarkContent(
+                content =
+                    Content(
+                        1L,
+                        Creator(1L, "채널명", ""),
+                        VideoData("콘텐츠 제목이 길면 ...으로 표시되는 것을 확인 ㅇㅇㅇ", "thumbnail", "2026-02-12"),
+                        City("대구"),
+                        true,
+                    ),
+                tripDuration = TripDuration(1, 2),
+                tripPlaceCount = 2,
+            )
+        BookmarkContentItem(
+            content = content,
+            onContentClick = {},
+            onRemoveBookmark = {},
+            modifier =
+                Modifier
+                    .padding(TuripTheme.spacing.large)
+                    .background(TuripTheme.colors.white),
+        )
+    }
+}
