--- android/app/src/main/java/com/on/turip/ui/compose/bookmarks/BookmarkContentUiEffect.kt
@@ -1,9 +1,0 @@
-package com.on.turip.ui.compose.bookmarks
-
-sealed interface BookmarkContentUiEffect {
-    data object NavigateToLogin : BookmarkContentUiEffect
-
-    data object ShowBookmarkRemoveFailed : BookmarkContentUiEffect
-
-    data object BookmarkRemoved : BookmarkContentUiEffect
-}
