+++ android/app/src/main/java/com/on/turip/ui/compose/mypage/util/MyPageMapper.kt
@@ -1,6 +1,6 @@
 package com.on.turip.ui.compose.mypage.util
 
-import com.on.turip.domain.accounts.Account
+import com.on.turip.domain.account.Account
 import com.on.turip.ui.compose.mypage.model.ProfileModel
 
 fun Account.toUiModel(): ProfileModel =
